filter lines by kakheti

+++ app/models/objects/line.rb
@@ -39,19 +39,21 @@ class Objects::Line
       idx1=descr.index(s1)+s1.length
       idx2=descr.index(s2)+s2.length
       regname=descr[idx1..-1].match(/<td>([^<])*<\/td>/)[0][4..-6].strip
-      direction=descr[idx2..-1].match(/<td>([^<])*<\/td>/)[0][4..-6].strip
-      region=Region.where(name:regname).first
-      region=Region.create(name:regname) if region.blank?
-      # end of description section
-      obj=Objects::Line.where(kmlid:id).first || Objects::Line.create(kmlid:id)
-      obj.direction=direction ; obj.region=region ; obj.name=name; obj.save
-      obj.points.destroy_all
-      coords.split(' ').each do |coord|
-        point=obj.points.new(line:obj)
-        point.set_coordinate(coord)
-        point.save
+      if regname == 'კახეთი'
+        direction=descr[idx2..-1].match(/<td>([^<])*<\/td>/)[0][4..-6].strip
+        region=Region.where(name:regname).first
+        region=Region.create(name:regname) if region.blank?
+        # end of description section
+        obj=Objects::Line.where(kmlid:id).first || Objects::Line.create(kmlid:id)
+        obj.direction=direction ; obj.region=region ; obj.name=name; obj.save
+        obj.points.destroy_all
+        coords.split(' ').each do |coord|
+          point=obj.points.new(line:obj)
+          point.set_coordinate(coord)
+          point.save
+        end
+        obj.calc_length!
       end
-      obj.calc_length!
     end
   end
